Update module setup for tests
